// src/models/configuration.model.ts
export interface CONFIGURATION  {
    configuration_id?:  number;
    configuration_title:  string;
    configuration_key:  string;
    configuration_value:  string;
    configuration_description:  string;
    configuration_group_id:  number;
    sort_order:  number;
    last_modified:  string;
    date_added:  string;
    use_function:  string;
    set_function:  string
}